+++ crates/op-rbuilder/src/limiter/mod.rs
@@ -1,45 +1,30 @@
-use std::{cmp::min, sync::Arc, time::Instant};
-
 use alloy_primitives::Address;
-use dashmap::DashMap;
 
-use crate::limiter::{args::GasLimiterArgs, error::GasLimitError, metrics::GasLimiterMetrics};
+use crate::limiter::{
+    args::GasLimiterArgs, error::GasLimitError, gas::GasLimiter, metrics::GasLimiterMetrics,
+};
 
 pub mod args;
 pub mod error;
+mod gas;
 mod metrics;
 
 #[derive(Debug, Clone)]
-pub struct AddressGasLimiter {
-    inner: Option<AddressGasLimiterInner>,
+pub struct AddressLimiter {
+    gas_limiter: Option<GasLimiter>,
 }
 
-#[derive(Debug, Clone)]
-struct AddressGasLimiterInner {
-    config: GasLimiterArgs,
-    // We don't need an Arc<Mutex<_>> here, we can get away with RefCell, but
-    // the reth PayloadBuilder trait needs this to be Send + Sync
-    address_buckets: Arc<DashMap<Address, TokenBucket>>,
-    metrics: GasLimiterMetrics,
-}
-
-#[derive(Debug, Clone)]
-struct TokenBucket {
-    capacity: u64,
-    available: u64,
-}
-
-impl AddressGasLimiter {
+impl AddressLimiter {
     pub fn new(config: GasLimiterArgs) -> Self {
         Self {
-            inner: AddressGasLimiterInner::try_new(config),
+            gas_limiter: GasLimiter::try_new(config),
         }
     }
 
     /// Check if there's enough gas for this address and consume it. Returns
     /// Ok(()) if there's enough otherwise returns an error.
     pub fn consume_gas(&self, address: Address, gas_requested: u64) -> Result<(), GasLimitError> {
-        if let Some(inner) = &self.inner {
+        if let Some(inner) = &self.gas_limiter {
             inner.consume_gas(address, gas_requested)
         } else {
             Ok(())
@@ -48,95 +33,8 @@ impl AddressGasLimiter {
 
     /// Should be called upon each new block. Refills buckets/Garbage collection
     pub fn refresh(&self, block_number: u64) {
-        if let Some(inner) = self.inner.as_ref() {
+        if let Some(inner) = self.gas_limiter.as_ref() {
             inner.refresh(block_number)
-        }
-    }
-}
-
-impl AddressGasLimiterInner {
-    fn try_new(config: GasLimiterArgs) -> Option<Self> {
-        if !config.gas_limiter_enabled {
-            return None;
-        }
-
-        Some(Self {
-            config,
-            address_buckets: Default::default(),
-            metrics: Default::default(),
-        })
-    }
-
-    fn consume_gas_inner(
-        &self,
-        address: Address,
-        gas_requested: u64,
-    ) -> Result<bool, GasLimitError> {
-        let mut created_new_bucket = false;
-        let mut bucket = self
-            .address_buckets
-            .entry(address)
-            // if we don't find a bucket we need to initialize a new one
-            .or_insert_with(|| {
-                created_new_bucket = true;
-                TokenBucket::new(self.config.max_gas_per_address)
-            });
-
-        if gas_requested > bucket.available {
-            return Err(GasLimitError::AddressLimitExceeded {
-                address,
-                requested: gas_requested,
-                available: bucket.available,
-            });
-        }
-
-        bucket.available -= gas_requested;
-
-        Ok(created_new_bucket)
-    }
-
-    fn consume_gas(&self, address: Address, gas_requested: u64) -> Result<(), GasLimitError> {
-        let start = Instant::now();
-        let result = self.consume_gas_inner(address, gas_requested);
-
-        self.metrics.record_gas_check(&result, start.elapsed());
-
-        result.map(|_| ())
-    }
-
-    fn refresh_inner(&self, block_number: u64) -> usize {
-        let active_addresses = self.address_buckets.len();
-
-        self.address_buckets.iter_mut().for_each(|mut bucket| {
-            bucket.available = min(
-                bucket.capacity,
-                bucket.available + self.config.refill_rate_per_block,
-            )
-        });
-
-        // Only clean up stale buckets every `cleanup_interval` blocks
-        if block_number.is_multiple_of(self.config.cleanup_interval) {
-            self.address_buckets
-                .retain(|_, bucket| bucket.available < bucket.capacity);
-        }
-
-        active_addresses - self.address_buckets.len()
-    }
-
-    fn refresh(&self, block_number: u64) {
-        let start = Instant::now();
-        let removed_addresses = self.refresh_inner(block_number);
-
-        self.metrics
-            .record_refresh(removed_addresses, start.elapsed());
-    }
-}
-
-impl TokenBucket {
-    fn new(capacity: u64) -> Self {
-        Self {
-            capacity,
-            available: capacity,
         }
     }
 }
@@ -162,7 +60,7 @@ mod tests {
     #[test]
     fn test_basic_refill() {
         let config = create_test_config(1000, 200, 10);
-        let limiter = AddressGasLimiter::new(config);
+        let limiter = AddressLimiter::new(config);
 
         // Consume all gas
         assert!(limiter.consume_gas(test_address(), 1000).is_ok());
@@ -177,7 +75,7 @@ mod tests {
     #[test]
     fn test_over_capacity_request() {
         let config = create_test_config(1000, 100, 10);
-        let limiter = AddressGasLimiter::new(config);
+        let limiter = AddressLimiter::new(config);
 
         // Request more than capacity should fail
         let result = limiter.consume_gas(test_address(), 1500);
@@ -195,7 +93,7 @@ mod tests {
     fn test_multiple_users() {
         // Simulate more realistic scenario
         let config = create_test_config(10_000_000, 1_000_000, 100); // 10M max, 1M refill
-        let limiter = AddressGasLimiter::new(config);
+        let limiter = AddressLimiter::new(config);
 
         let searcher1 = Address::from([0x1; 20]);
         let searcher2 = Address::from([0x2; 20]);
@@ -219,44 +117,5 @@ mod tests {
         assert!(limiter.consume_gas(searcher1, 1_000_000).is_ok()); // Had 9.5M + 1M refill, now 9.5M
         assert!(limiter.consume_gas(searcher2, 1_000_000).is_ok()); // Had 9.25M + 1M refill, now 9.25M
         assert!(limiter.consume_gas(attacker, 1_000_000).is_ok()); // Had 5M + 1M refill, now 5M
-    }
-
-    #[test]
-    fn test_bucket_cleanup() {
-        // Test that unused buckets get cleaned up properly
-        let config = create_test_config(1000, 1000, 10);
-        let limiter = AddressGasLimiter::new(config);
-
-        let addr1 = Address::from([0x1; 20]);
-        let addr2 = Address::from([0x2; 20]);
-
-        // Create buckets for both
-        assert!(limiter.consume_gas(addr1, 100).is_ok());
-        assert!(limiter.consume_gas(addr2, 100).is_ok());
-
-        let inner = limiter.inner.as_ref().unwrap();
-        assert_eq!(inner.address_buckets.len(), 2);
-
-        // Refill for several blocks - addr1 stays at full capacity (unused)
-        // but addr2 continues to be used
-        for block in 1..=10 {
-            limiter.refresh(block);
-
-            if block > 1 {
-                // addr1 is now full and unused
-                // addr2 continues to use gas
-                assert!(limiter.consume_gas(addr2, 100).is_ok());
-            }
-        }
-
-        // After cleanup at block 10 (multiple of 5), addr1 should be removed
-        // because it's at full capacity (unused), while addr2 remains
-        assert_eq!(
-            inner.address_buckets.len(),
-            1,
-            "Unused bucket (addr1) should have been cleaned up"
-        );
-        assert!(inner.address_buckets.contains_key(&addr2));
-        assert!(!inner.address_buckets.contains_key(&addr1));
     }
 }
